// src/pages/hr/empManagement/types/empManagementTypes.ts
export interface EmpInfoEntity {
  empCode?: string | undefined | null;
  empName?: string | undefined | null;
  email?: string | undefined | null;
  gender?: string | undefined | null;
  address?: string | undefined | null;
  birthDate?: string | undefined | null;
  mobileNumber?: string | undefined | null;
  detailAddress?: string | undefined | null;
  postNumber?: string | undefined | null;
  residentId?: string | undefined;
  deptCode?: string | undefined;
  educationInfoList?: EducationInfoList[];
}
export interface EducationInfoList {
  lastSchool?: string | undefined | null;
  schoolName?: string | undefined | null;
  major?: string | undefined | null;
  entranceDate?: string | undefined | null;
  graduateDate?: string | undefined | null;
}